+++ src/isotools/_transcriptome_stats.py
@@ -6,10 +6,9 @@ import logging
 import numpy as np
 import pandas as pd
 import itertools
-from scipy.stats import chi2_contingency, fisher_exact
 
 # from .decorators import deprecated, debug, experimental
-from ._utils import _filter_function, _corrected_log2OR
+from ._utils import _filter_function
 
 logger = logging.getLogger('isotools')
 
@@ -738,76 +737,8 @@ def rarefaction(self, groups=None, fractions=20, min_coverage=2, tr_filter={}):
     return pd.DataFrame(curves, index=fractions), total
 
 
-def pairwise_event_test(e1, e2, coverage, test="chi2"):
-    '''
-    performs an independence test among the joint frequencies of two events
-
-    :param e1: event obtained from .find_splice_bubbles()
-    :param e2: event obtained from .find_splice_bubbles()
-    :param transcriptome: the Transcriptome object on which the events were computed
-    :param coverage: list of counts per transcript
-    :param min_dist: minimum distance (in nucleotides) between the two Alternative Splicing Events for the pair to be tested
-    :type min_dist: int
-    :param test: test to be performed. One of ("chi2", "fisher")
-    :type test: str
-    :param sg: segment graph
-    '''
-
-    con_tab = np.zeros((2, 2), dtype=int)
-    tr_ID_tab = np.zeros((2, 2), dtype=object)
-
-    for m, n in itertools.product(range(2), range(2)):
-        tr_IDs = set(e1[m]) & set(e2[n])
-        cov_tr_IDs = [(coverage[trid], trid) for trid in tr_IDs]  # get coverage and tr_ID
-        cov_tr_IDs.sort(reverse=True)  # sort by coverage
-        tr_ID_tab[n, m] = [trid for _, trid in cov_tr_IDs]
-
-        c_mn = sum([cov for cov, _ in cov_tr_IDs])
-        con_tab[n, m] = c_mn
-
-    if test == 'chi2':
-        test_fun = chi2_contingency
-    elif test == 'fisher':
-        test_fun = fisher_exact
-
-    test_res = test_fun(con_tab+.01)  # add some small value (TODO: is this also for fisher test?)
-
-    priA_priB_trID, altA_altB_trID = tr_ID_tab[0, 0], tr_ID_tab[1, 1]
-    priA_altB_trID, altA_priB_trID = tr_ID_tab[1, 0], tr_ID_tab[0, 1]
-    priA_priB, altA_altB = con_tab[0, 0], con_tab[1, 1]
-    priA_altB, altA_priB = con_tab[1, 0], con_tab[0, 1]
-    p_value = test_res[1]
-    test_stat = test_res[0]
-    log2OR = _corrected_log2OR(con_tab)
-
-    # logOR is a measure of the effect size. coordination between the events is either positive or negative.
-
-    return_tuple = (p_value, test_stat, log2OR, priA_priB, priA_altB, altA_priB, altA_altB,
-                    priA_priB_trID, priA_altB_trID, altA_priB_trID, altA_altB_trID)
-
-    return return_tuple
-
-
-def precompute_events_dict(self, event_type=("ES", "5AS", "3AS", "IR", "ME"), min_cov=100, region=None, progress_bar=True):
-    '''
-    Precomputes the evvents_dict, i.e. a dictionary of splice bubbles. Each key is a gene and each value is the splice bubbles
-    object corresponding to that gene.
-    :param region: The region to be considered. Either a string "chr:start-end", or a tuple (chr,start,end). Start and end is optional.
-
-    '''
-
-    events_dict = {}
-
-    for g in self.iter_genes(region=region, progress_bar=progress_bar):
-        sg = g.segment_graph
-        events = sg.find_splice_bubbles(types=event_type)
-        events_dict[g.id] = [e for e in events if g.coverage.sum(axis=0)[e[0]+e[1]].sum() >= min_cov]
-
-    return events_dict
-
-
-def coordination_test(self, samples=None, test="chi2", min_dist=1, min_total=100, min_alt_fraction=.1, min_cov_pair=100,
-                      events_dict={}, event_type=("ES", "5AS", "3AS", "IR", "ME"), query=None, region=None, method="fdr_bh",
+def coordination_test(self, samples=None, test="fisher", min_dist=1, min_total=100, min_alt_fraction=.1,
+                      events_dict={}, event_type=("ES", "5AS", "3AS", "IR", "ME"), query=None, region=None, padj_method="fdr_bh",
                       progress_bar=True):
     '''
     Performs gene_coordination_test on all genes.
@@ -830,7 +761,8 @@ def coordination_test(self, samples=None, test="chi2", min_dist=1, min_total=100
     Default is ("ES", "5AS", "3AS", "IR", "ME")
     :param query: If provided, query string is evaluated on all genes for filtering
     :param region: The region to be considered. Either a string "chr:start-end", or a tuple (chr,start,end). Start and end is optional.
-    :param method: The multiple tet adjustment method. Any value allowed by statsmodels.stats.multitest.multipletests (std: Benjamini-Hochberg)
+    :param padj_method: The multiple test adjustment method.
+        Any value allowed by statsmodels.stats.multitest.multipletests (default: Benjamini-Hochberg)
 
     :return: a Pandas dataframe, where each column corresponds to the p_values, the statistics
     (the chi squared statistic if the chi squared test is used and the odds-ratio if the Fisher
@@ -850,7 +782,7 @@ def coordination_test(self, samples=None, test="chi2", min_dist=1, min_total=100
 
         try:
             next_test_res = g.coordination_test(test=test, samples=samples, min_dist=min_dist, min_total=min_total,
-                                                min_alt_fraction=min_alt_fraction, min_cov_pair=min_cov_pair,
+                                                min_alt_fraction=min_alt_fraction,
                                                 events=events_dict.get(g.id), event_type=event_type)
             test_res.extend(next_test_res)
 
@@ -858,14 +790,14 @@ def coordination_test(self, samples=None, test="chi2", min_dist=1, min_total=100
             logger.error(f"\nError encounter on {print(g)} {g.id}   :  {g.name}.")
             raise
 
-    col_names = ("gene_id", "gene_name", "strand", "ase1_type", "ase2_type", "ase1_start", "ase1_end",
-                 "ase2_start", "ase2_end", "p_value", "stat", "log2OR", "priA_priB", "priA_altB", "altA_priB",
+    col_names = ("gene_id", "gene_name", "strand", "eventA_type", "eventB_type", "eventA_start", "evemtA_end",
+                 "eventB_start", "eventB_end", "pvalue", "stat", "log2OR", "dcPSI_AB", "dcPSI_BA", "priA_priB", "priA_altB", "altA_priB",
                  "altA_altB", "priA_priB_trID", "priA_altB_trID", "altA_priB_trID", "altA_altB_trID")
 
     res = pd.DataFrame(test_res, columns=col_names)
 
-    adj_p_value = multi.multipletests(res.p_value, method=method)[1]
+    adj_p_value = multi.multipletests(res.pvalue, method=padj_method)[1]
 
-    res.insert(10, "adj_p_value", adj_p_value)
+    res.insert(10, "padj", adj_p_value)
 
     return res